Moved the spells screen to the manage spells screen and reversed the spells screen.
