+++ BeyondSW5E/src/screens/ManageSpellsScreen.js
@@ -1,14 +1,43 @@
-import React from 'react'
-import { Text, View, StyleSheet } from 'react-native'
+import React, { useState, useEffect } from 'react'
+import { Text, View, StyleSheet, Button } from 'react-native'
+import swapi from '../api/swapi'
+import PowerCardList from '../components/PowerCardList'
 
 const ManageSpellsScreen = () => {
+    const [power, setPower] = useState([])
+
+    const searchApi = async () => {
+        const response = await swapi.get('/power');
+        setPower(response.data)
+    }
+
+    useEffect(() => { searchApi() }, [])
+
     return (
-        <View>
-            <Text>Manage Force/Tech Powers Screen</Text>
+        <View style = { styles.screenContainer }>
+            <Text style = { styles.headerStyle }>Powers Screen</Text>
+            <Text>Displaying {power.length} powers.</Text>
+            {
+                power.length === 0
+                ? <Text>Equipment is Empty</Text>
+                : <PowerCardList
+                      powers = { power }
+                  />
+            }
         </View>
     )
 }
 
-const styles = StyleSheet.create({})
+const styles = StyleSheet.create({
+    screenContainer: {
+        alignItems: 'center',
+        flex: 1
+    },
+    headerStyle: {
+        fontSize: 24,
+        fontWeight: 'bold',
+        alignSelf: 'center'
+    }
+})
 
 export default ManageSpellsScreen
